Only update size when values are valid
In some environments the ioctl call for window size will return 0x0,
for instance in serial console setups. This causes the input to be
truncated making it hard for users to see what they are typing. This
performs a check to see if either width or height has a non-zero value
before updating the size.