editline: add live, token-based syntax highlighting
This introduces support for live, "as-you-type" syntax highlighting
to improve readability for structured text formats like SQL.

The main change is a new token-based rendering path. A new public
`highlight` package provides the core `Highlighter` interface, and the
internal `textarea` component has been refactored to be token-aware.

The new feature is exposed via a `SetHighlighter()` method on the
`bubbline.Editor` and `editline.Model`.