Update test expectations based on latest dep upgrade
This patch updates the expected test outputs.

This reveals two user-visible changes in the Bubbline editor:

1. when there are fewer lines of input than display lines in the
   textarea, the remaining display lines are now filled with
   whitespace (instead of "~" as previously).
   This is change 620ca7dfa9fee23c2b01b153bdd3a913591a42d7 in the
   bubbles project (issue charmbracelet/bubbles#510).

2. the keyboard hints under the prompt are not truncated any more when
   the textarea width is configured to be narrower than the hints.
   The specific cause of this change is unclear, but it might
   be change 48dffdd97e4ff8cb3fb16e6414bc8b6d8ae86a4c in the bubbles project.

We now also need to explicitly set the color output profile in
lipgloss for tests, because since ~2023 termenv auto-detects the color
profile based on the output stream. In tests, that autodetection
disables styling by default.